# apps/openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "My Catatan Kuliah",
    "version": "1.0.0",
    "description": "My Catatan Kuliah API"
  },
  "servers": [
    {
      "url": "/",
      "description": "Default server",
      "variables": {}
    }
  ],
  "paths": {
    "/api/assignments": {
      "post": {
        "operationId": "assignmentsServiceCreateAssignment",
        "description": "Create a new assignment with the specified details.",
        "parameters": [],
        "requestBody": {
          "description": "The details of the assignment to create",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignmentsCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The request has succeeded and a new resource has been created as a result.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentsCreateResponse"
                }
              }
            }
          },
          "400": {
            "description": "The server could not understand the request due to invalid syntax.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError400"
                }
              }
            }
          },
          "401": {
            "description": "Access is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError401"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError500"
                }
              }
            }
          }
        },
        "tags": [
          "Assignments"
        ]
      },
      "get": {
        "operationId": "assignmentsServiceListAssignments",
        "description": "List all assignments accessible to the authenticated user.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentsListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Access is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError401"
                }
              }
            }
          },
          "404": {
            "description": "The server cannot find the requested resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError404"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError500"
                }
              }
            }
          }
        },
        "tags": [
          "Assignments"
        ]
      }
    },
    "/api/assignments/{assignmentId}": {
      "get": {
        "operationId": "assignmentsServiceGetAssignment",
        "description": "Get details of a specific assignment by its ID.",
        "parameters": [
          {
            "name": "assignmentId",
            "in": "path",
            "description": "The unique identifier of the assignment",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-oapi-codegen-extra-tags": {
              "validate": "required"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentsDetailResponse"
                }
              }
            }
          },
          "401": {
            "description": "Access is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError401"
                }
              }
            }
          },
          "404": {
            "description": "The server cannot find the requested resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError404"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError500"
                }
              }
            }
          }
        },
        "tags": [
          "Assignments"
        ]
      },
      "delete": {
        "operationId": "assignmentsServiceDeleteAssignment",
        "description": "Delete a specific assignment by its ID.",
        "parameters": [
          {
            "name": "assignmentId",
            "in": "path",
            "description": "The unique identifier of the assignment",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-oapi-codegen-extra-tags": {
              "validate": "required"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "There is no content to send for this request, but the headers may be useful. "
          },
          "401": {
            "description": "Access is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError401"
                }
              }
            }
          },
          "404": {
            "description": "The server cannot find the requested resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError404"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError500"
                }
              }
            }
          }
        },
        "tags": [
          "Assignments"
        ]
      },
      "patch": {
        "operationId": "assignmentsServiceUpdateAssignment",
        "description": "Update details of a specific assignment by its ID.",
        "parameters": [
          {
            "name": "assignmentId",
            "in": "path",
            "description": "The unique identifier of the assignment",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-oapi-codegen-extra-tags": {
              "validate": "required"
            }
          }
        ],
        "requestBody": {
          "description": "The details to update for the assignment",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignmentsUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentsUpdateResponse"
                }
              }
            }
          },
          "400": {
            "description": "The server could not understand the request due to invalid syntax.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError400"
                }
              }
            }
          },
          "401": {
            "description": "Access is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError401"
                }
              }
            }
          },
          "404": {
            "description": "The server cannot find the requested resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError404"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError500"
                }
              }
            }
          }
        },
        "tags": [
          "Assignments"
        ]
      }
    },
...
    "/api/courses": {
      "post": {
        "operationId": "coursesServiceCreateCourse",
        "description": "Create a new course with the specified details.",
        "parameters": [],
        "requestBody": {
          "description": "The details of the course to create",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CoursesCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The request has succeeded and a new resource has been created as a result.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoursesCreateResponse"
                }
              }
            }
          },
          "400": {
            "description": "The server could not understand the request due to invalid syntax.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError400"
                }
              }
            }
          },
          "401": {
            "description": "Access is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError401"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError500"
                }
              }
            }
          }
        },
        "tags": [
          "Courses"
        ]
      },
      "get": {
        "operationId": "coursesServiceListCourses",
        "description": "List all courses accessible to the authenticated user.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoursesListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Access is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError401"
                }
              }
            }
          },
          "404": {
            "description": "The server cannot find the requested resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError404"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError500"
                }
              }
            }
          }
        },
        "tags": [
          "Courses"
        ]
      }
    },
    "/api/courses/{courseId}": {
      "get": {
        "operationId": "coursesServiceGetCourse",
        "description": "Get details of a specific course by its ID.",
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "description": "The unique identifier of the course",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-oapi-codegen-extra-tags": {
              "validate": "required"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoursesResponse"
                }
              }
            }
          },
          "401": {
            "description": "Access is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError401"
                }
              }
            }
          },
          "404": {
            "description": "The server cannot find the requested resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError404"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError500"
                }
              }
            }
          }
        },
        "tags": [
          "Courses"
        ]
      },
      "delete": {
        "operationId": "coursesServiceDeleteCourse",
        "description": "Delete a specific course by its ID.",
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "description": "The unique identifier of the course",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-oapi-codegen-extra-tags": {
              "validate": "required"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "There is no content to send for this request, but the headers may be useful. "
          },
          "401": {
            "description": "Access is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError401"
                }
              }
            }
          },
          "404": {
            "description": "The server cannot find the requested resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError404"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError500"
                }
              }
            }
          }
        },
        "tags": [
          "Courses"
        ]
      },
      "patch": {
        "operationId": "coursesServiceUpdateCourse",
        "description": "Update details of a specific course by its ID.",
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "description": "The unique identifier of the course",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-oapi-codegen-extra-tags": {
              "validate": "required"
            }
          }
        ],
        "requestBody": {
          "description": "The details to update for the course",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CoursesUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoursesResponse"
                }
              }
            }
          },
          "400": {
            "description": "The server could not understand the request due to invalid syntax.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError400"
                }
              }
            }
          },
          "401": {
            "description": "Access is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError401"
                }
              }
            }
          },
          "404": {
            "description": "The server cannot find the requested resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError404"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError500"
                }
              }
            }
          }
        },
        "tags": [
          "Courses"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AssignmentsAssignmentStatus": {
        "type": "string",
        "enum": [
          "Todo",
          "InProgress",
          "Done"
        ],
        "x-oapi-codegen-extra-tags": {
          "validate": "omitempty,oneof=Todo InProgress Done"
        }
      },
      "AssignmentsCreateRequest": {
        "type": "object",
        "properties": {
          "course_id": {
            "type": "string",
            "x-oapi-codegen-extra-tags": {
              "validate": "required"
            }
          },
          "title": {
            "type": "string",
            "x-oapi-codegen-extra-tags": {
              "validate": "required"
            }
          },
          "status": {
            "$ref": "#/components/schemas/AssignmentsAssignmentStatus"
          },
          "position": {
            "type": "integer",
            "x-oapi-codegen-extra-tags": {
              "validate": "required"
            }
          },
          "due_date": {
            "type": "string",
            "format": "date-time",
            "x-oapi-codegen-extra-tags": {
              "validate": "required,datetime=2006-01-02T15:04:05Z07:00"
            }
          },
          "content": {
            "type": "string",
            "x-oapi-codegen-extra-tags": {
              "validate": "omitempty"
            }
          }
        },
        "required": [
          "course_id",
          "title",
          "status",
          "position",
          "due_date"
        ]
      },
      "AssignmentsCreateResponse": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/AssignmentsDetailResponse"
          }
        ]
      },
      "AssignmentsDetailResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "readOnly": true,
            "x-oapi-codegen-extra-tags": {
              "validate": "required"
            }
          },
          "course_id": {
            "type": "string",
            "x-oapi-codegen-extra-tags": {
              "validate": "required"
            }
          },
          "title": {
            "type": "string",
            "x-oapi-codegen-extra-tags": {
              "validate": "required"
            }
          },
          "status": {
            "$ref": "#/components/schemas/AssignmentsAssignmentStatus"
          },
          "position": {
            "type": "integer",
            "x-oapi-codegen-extra-tags": {
              "validate": "required"
            }
          },
          "due_date": {
            "type": "string",
            "format": "date-time",
            "x-oapi-codegen-extra-tags": {
              "validate": "required,datetime=2006-01-02T15:04:05Z07:00"
            }
          },
          "created_by": {
            "type": "string",
            "readOnly": true,
            "x-oapi-codegen-extra-tags": {
              "validate": "required"
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "readOnly": true,
            "x-oapi-codegen-extra-tags": {
              "validate": "required,datetime=2006-01-02T15:04:05Z07:00"
            }
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "readOnly": true,
            "x-oapi-codegen-extra-tags": {
              "validate": "required,datetime=2006-01-02T15:04:05Z07:00"
            }
          },
          "content": {
            "type": "string",
            "x-oapi-codegen-extra-tags": {
              "validate": "omitempty"
            }
          }
        },
        "required": [
          "id",
          "course_id",
          "title",
          "status",
          "position",
          "due_date",
          "created_by",
          "created_at",
          "updated_at"
        ]
      },
      "AssignmentsListResponse": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/AssignmentsResponse"
        },
        "x-oapi-codegen-extra-tags": {
          "validate": "omitempty"
        }
      },
...
      "AssignmentsUpdateRequest": {
        "description": "The template for adding optional properties.",
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "x-oapi-codegen-extra-tags": {
              "validate": "omitempty"
            }
          },
          "status": {
            "$ref": "#/components/schemas/AssignmentsAssignmentStatus"
          },
          "position": {
            "type": "integer",
            "x-oapi-codegen-extra-tags": {
              "validate": "omitempty"
            }
          },
          "due_date": {
            "type": "string",
            "format": "date-time",
            "x-oapi-codegen-extra-tags": {
              "validate": "omitempty,datetime=2006-01-02T15:04:05Z07:00"
            }
          },
          "content": {
            "type": "string",
            "x-oapi-codegen-extra-tags": {
              "validate": "omitempty"
            }
          }
        }
      },
      "AssignmentsUpdateResponse": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/AssignmentsDetailResponse"
          }
        ]
      },
...
      "CommonErrorError400": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/CommonErrorError"
          }
        ]
      },
      "CommonErrorError401": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/CommonErrorError"
          }
        ]
      },
      "CommonErrorError404": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/CommonErrorError"
          }
        ]
      },
      "CommonErrorError500": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/CommonErrorError"
          }
        ]
      },
      "CoursesCreateRequest": {
        "type": "object",
        "properties": {
          "workspace_id": {
            "type": "string",
            "x-oapi-codegen-extra-tags": {
              "validate": "required"
            }
          },
          "title": {
            "type": "string",
            "x-oapi-codegen-extra-tags": {
              "validate": "required"
            }
          },
          "instructor": {
            "type": "string",
            "x-oapi-codegen-extra-tags": {
              "validate": "required"
            }
          },
          "credits": {
            "type": "integer",
            "x-oapi-codegen-extra-tags": {
              "validate": "required"
            }
          }
        },
        "required": [
          "workspace_id",
          "title",
          "instructor",
          "credits"
        ]
      },
      "CoursesCreateResponse": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/CoursesResponse"
          }
        ]
      },
      "CoursesListResponse": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/CoursesResponse"
        },
        "x-oapi-codegen-extra-tags": {
          "validate": "omitempty"
        }
      },
      "CoursesResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "readOnly": true,
            "x-oapi-codegen-extra-tags": {
              "validate": "required"
            }
          },
          "workspace_id": {
            "type": "string",
            "x-oapi-codegen-extra-tags": {
              "validate": "required"
            }
          },
          "title": {
            "type": "string",
            "x-oapi-codegen-extra-tags": {
              "validate": "required"
            }
          },
          "instructor": {
            "type": "string",
            "x-oapi-codegen-extra-tags": {
              "validate": "required"
            }
          },
          "credits": {
            "type": "integer",
            "x-oapi-codegen-extra-tags": {
              "validate": "required"
            }
          },
          "created_by": {
            "type": "string",
            "readOnly": true,
            "x-oapi-codegen-extra-tags": {
              "validate": "required"
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "readOnly": true,
            "x-oapi-codegen-extra-tags": {
              "validate": "required,datetime=2006-01-02T15:04:05Z07:00"
            }
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "readOnly": true,
            "x-oapi-codegen-extra-tags": {
              "validate": "required,datetime=2006-01-02T15:04:05Z07:00"
            }
          }
        },
        "required": [
          "id",
          "workspace_id",
          "title",
          "instructor",
          "credits",
          "created_by",
          "created_at",
          "updated_at"
        ]
      },
      "CoursesUpdateRequest": {
        "description": "The template for adding optional properties.",
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "x-oapi-codegen-extra-tags": {
              "validate": "omitempty"
            }
          },
          "instructor": {
            "type": "string",
            "x-oapi-codegen-extra-tags": {
              "validate": "omitempty"
            }
          },
          "credits": {
            "type": "integer",
            "x-oapi-codegen-extra-tags": {
              "validate": "omitempty"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    }
  },
...
  "security": [
    {
      "ApiKeyAuth": []
    }
  ]
}